Throw error if we are trying to build in master

+++ rel-eng/build.sh
@@ -4,9 +4,14 @@ set -x
 set -e
 
 name=$( grep '^Name:' rel-eng/satellite-performance.spec | sed 's/^Name:\s*\([a-z-]\+\)\s*$/\1/' )
-version=$( grep '^Version:' rel-eng/satellite-performance.spec | sed 's/^Version:\s*\([0-9.]\+\)\s*$/\1/' )
+version=$( grep '^Version:' rel-eng/satellite-performance.spec | sed 's/^Version:\s*\([0-9a-z.]\+\)\s*$/\1/' )
 directory="$name-$version"
 archive="$directory.tar.gz"
+
+if [ "$version" = 'master' ]; then
+  echo "ERROR: Make sure you are in release branch and update version number in rel-eng/satellite-performance.spec first" >&2
+  exit 1
+fi
 
 # Prepare directory
 rm -rf /tmp/$directory
